third_party/blink/web_tests/fast/events/wheel/wheelevent-in-horizontal-scrollbar-in-rtl.html

<html dir=rtl>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/gesture-util.js"></script>
<script src="../../../resources/percent-based-util.js"></script>
<script>
  const pixelsToScroll1 = 10;
  const { x: pixelsToScrollX1, y: pixelsToScrollY1 } =
      calculatePixelsToScroll(document.scrollingElement, pixelsToScroll1,
                              pixelsToScroll1);
  const pixelsToScroll2 = 30;
  const { x: pixelsToScrollX2, y: pixelsToScrollY2 } =
      calculatePixelsToScroll(document.scrollingElement, pixelsToScroll2,
                              pixelsToScroll2);
  const testScrolls = [
    { distance: pixelsToScrollX1, expectedX: -90, expectedY: 100,
      direction: 'right' },
    { distance: pixelsToScrollX2, expectedX: -120, expectedY: 100,
      direction: 'left' },
    { distance: pixelsToScrollY1, expectedX: -120, expectedY: 110,
      direction: 'down' },
    { distance: pixelsToScrollY2, expectedX: -120, expectedY: 80,
      direction: 'up' }
  ];
  let currentTest = -1;

  function startNextTestCase() {
    currentTest++;
    if (currentTest >= testScrolls.length)
      return;

    let testCase = testScrolls[currentTest];
    promise_test(async () => {
      await mouseMoveTo(100, 100);
      await gestureScroll(() => {
        return smoothScroll(
            testCase.distance, 100, 100,
            GestureSourceType.MOUSE_INPUT, testCase.direction, SPEED_INSTANT,
            !isPercentBasedScrollingEnabled());
      }, document.scrollingElement);
      assert_approx_equals(window.scrollX, testCase.expectedX, 1);
      assert_approx_equals(window.scrollY, testCase.expectedY, 1);
      startNextTestCase();
    }, 'scroll ' + testCase.direction);
  }

  window.onload = async function() {
    await waitForScrollReset(document.scrollingElement, -100, 100);
    startNextTestCase();
  }
</script>

<body style="margin:0">
  <div id="div" style="whitespace:nowrap; width: 4000px; height: 1000px; border: 1px solid red;">Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum."</div>
  <div id="console"></div>
</body>
</html>
